import ts compatibility
