+++ src/store/schemaStore.js
@@ -1,3 +1,335 @@
+const departmentSchema = {
+  departments: {
+    dept_no: {
+      Name: 'dept_no',
+      Value: null,
+      data_type: 'char',
+      TableName: 'departments',
+      References: [],
+      IsPrimaryKey: true,
+      IsForeignKey: false,
+      additional_constraints: 'PRIMARY KEY',
+      field_name: 'dept_no',
+    },
+    dept_name: {
+      Name: 'dept_name',
+      Value: null,
+      data_type: 'varchar',
+      TableName: 'departments',
+      References: [],
+      IsPrimaryKey: false,
+      IsForeignKey: false,
+      additional_constraints: 'NOT NULL',
+      field_name: 'dept_name',
+    },
+  },
+  dept_emp: {
+    emp_no: {
+      Name: 'emp_no',
+      Value: null,
+      data_type: 'int',
+      TableName: 'dept_emp',
+      References: [
+        {
+          PrimaryKeyName: 'emp_no',
+          PrimaryKeyTableName: 'employees',
+          ReferencesPropertyName: 'emp_no',
+          ReferencesTableName: 'dept_emp',
+          IsDestination: false,
+          constrainName: 'dept_emp_ibfk_1',
+        },
+      ],
+      IsPrimaryKey: true,
+      IsForeignKey: true,
+      additional_constraints: 'PRIMARY KEY',
+      field_name: 'emp_no',
+    },
+    dept_no: {
+      Name: 'dept_no',
+      Value: null,
+      data_type: 'char',
+      TableName: 'dept_emp',
+      References: [
+        {
+          PrimaryKeyName: 'dept_no',
+          PrimaryKeyTableName: 'departments',
+          ReferencesPropertyName: 'dept_no',
+          ReferencesTableName: 'dept_emp',
+          IsDestination: false,
+          constrainName: 'dept_emp_ibfk_2',
+        },
+      ],
+      IsPrimaryKey: true,
+      IsForeignKey: true,
+      additional_constraints: 'PRIMARY KEY',
+      field_name: 'dept_no',
+    },
+    from_date: {
+      Name: 'from_date',
+      Value: null,
+      data_type: 'date',
+      TableName: 'dept_emp',
+      References: [],
+      IsPrimaryKey: false,
+      IsForeignKey: false,
+      additional_constraints: 'NOT NULL',
+      field_name: 'from_date',
+    },
+    to_date: {
+      Name: 'to_date',
+      Value: null,
+      data_type: 'date',
+      TableName: 'dept_emp',
+      References: [],
+      IsPrimaryKey: false,
+      IsForeignKey: false,
+      additional_constraints: 'NOT NULL',
+      field_name: 'to_date',
+    },
+  },
+  dept_manager: {
+    emp_no: {
+      Name: 'emp_no',
+      Value: null,
+      data_type: 'int',
+      TableName: 'dept_manager',
+      References: [
+        {
+          PrimaryKeyName: 'emp_no',
+          PrimaryKeyTableName: 'employees',
+          ReferencesPropertyName: 'emp_no',
+          ReferencesTableName: 'dept_manager',
+          IsDestination: false,
+          constrainName: 'dept_manager_ibfk_1',
+        },
+      ],
+      IsPrimaryKey: true,
+      IsForeignKey: true,
+      additional_constraints: 'PRIMARY KEY',
+      field_name: 'emp_no',
+    },
+    dept_no: {
+      Name: 'dept_no',
+      Value: null,
+      data_type: 'char',
+      TableName: 'dept_manager',
+      References: [
+        {
+          PrimaryKeyName: 'dept_no',
+          PrimaryKeyTableName: 'departments',
+          ReferencesPropertyName: 'dept_no',
+          ReferencesTableName: 'dept_manager',
+          IsDestination: false,
+          constrainName: 'dept_manager_ibfk_2',
+        },
+      ],
+      IsPrimaryKey: true,
+      IsForeignKey: true,
+      additional_constraints: 'PRIMARY KEY',
+      field_name: 'dept_no',
+    },
+    from_date: {
+      Name: 'from_date',
+      Value: null,
+      data_type: 'date',
+      TableName: 'dept_manager',
+      References: [],
+      IsPrimaryKey: false,
+      IsForeignKey: false,
+      additional_constraints: 'NOT NULL',
+      field_name: 'from_date',
+    },
+    to_date: {
+      Name: 'to_date',
+      Value: null,
+      data_type: 'date',
+      TableName: 'dept_manager',
+      References: [],
+      IsPrimaryKey: false,
+      IsForeignKey: false,
+      additional_constraints: 'NOT NULL',
+      field_name: 'to_date',
+    },
+  },
+  employees: {
+    emp_no: {
+      Name: 'emp_no',
+      Value: null,
+      data_type: 'int',
+      TableName: 'employees',
+      References: [],
+      IsPrimaryKey: true,
+      IsForeignKey: false,
+      additional_constraints: 'PRIMARY KEY',
+      field_name: 'emp_no',
+    },
+    birth_date: {
+      Name: 'birth_date',
+      Value: null,
+      data_type: 'date',
+      TableName: 'employees',
+      References: [],
+      IsPrimaryKey: false,
+      IsForeignKey: false,
+      additional_constraints: 'NOT NULL',
+      field_name: 'birth_date',
+    },
+    first_name: {
+      Name: 'first_name',
+      Value: null,
+      data_type: 'varchar',
+      TableName: 'employees',
+      References: [],
+      IsPrimaryKey: false,
+      IsForeignKey: false,
+      additional_constraints: 'NOT NULL',
+      field_name: 'first_name',
+    },
+    last_name: {
+      Name: 'last_name',
+      Value: null,
+      data_type: 'varchar',
+      TableName: 'employees',
+      References: [],
+      IsPrimaryKey: false,
+      IsForeignKey: false,
+      additional_constraints: 'NOT NULL',
+      field_name: 'last_name',
+    },
+    gender: {
+      Name: 'gender',
+      Value: null,
+      data_type: 'enum',
+      TableName: 'employees',
+      References: [],
+      IsPrimaryKey: false,
+      IsForeignKey: false,
+      additional_constraints: 'NOT NULL',
+      field_name: 'gender',
+    },
+    hire_date: {
+      Name: 'hire_date',
+      Value: null,
+      data_type: 'date',
+      TableName: 'employees',
+      References: [],
+      IsPrimaryKey: false,
+      IsForeignKey: false,
+      additional_constraints: 'NOT NULL',
+      field_name: 'hire_date',
+    },
+  },
+  salaries: {
+    emp_no: {
+      Name: 'emp_no',
+      Value: null,
+      data_type: 'int',
+      TableName: 'salaries',
+      References: [
+        {
+          PrimaryKeyName: 'emp_no',
+          PrimaryKeyTableName: 'employees',
+          ReferencesPropertyName: 'emp_no',
+          ReferencesTableName: 'salaries',
+          IsDestination: false,
+          constrainName: 'salaries_ibfk_1',
+        },
+      ],
+      IsPrimaryKey: true,
+      IsForeignKey: true,
+      additional_constraints: 'PRIMARY KEY',
+      field_name: 'emp_no',
+    },
+    salary: {
+      Name: 'salary',
+      Value: null,
+      data_type: 'int',
+      TableName: 'salaries',
+      References: [],
+      IsPrimaryKey: false,
+      IsForeignKey: false,
+      additional_constraints: 'NOT NULL',
+      field_name: 'salary',
+    },
+    from_date: {
+      Name: 'from_date',
+      Value: null,
+      data_type: 'date',
+      TableName: 'salaries',
+      References: [],
+      IsPrimaryKey: true,
+      IsForeignKey: false,
+      additional_constraints: 'PRIMARY KEY',
+      field_name: 'from_date',
+    },
+    to_date: {
+      Name: 'to_date',
+      Value: null,
+      data_type: 'date',
+      TableName: 'salaries',
+      References: [],
+      IsPrimaryKey: false,
+      IsForeignKey: false,
+      additional_constraints: 'NOT NULL',
+      field_name: 'to_date',
+    },
+  },
+  titles: {
+    emp_no: {
+      Name: 'emp_no',
+      Value: null,
+      data_type: 'int',
+      TableName: 'titles',
+      References: [
+        {
+          PrimaryKeyName: 'emp_no',
+          PrimaryKeyTableName: 'employees',
+          ReferencesPropertyName: 'emp_no',
+          ReferencesTableName: 'titles',
+          IsDestination: false,
+          constrainName: 'titles_ibfk_1',
+        },
+      ],
+      IsPrimaryKey: true,
+      IsForeignKey: true,
+      additional_constraints: 'PRIMARY KEY',
+      field_name: 'emp_no',
+    },
+    title: {
+      Name: 'title',
+      Value: null,
+      data_type: 'varchar',
+      TableName: 'titles',
+      References: [],
+      IsPrimaryKey: true,
+      IsForeignKey: false,
+      additional_constraints: 'PRIMARY KEY',
+      field_name: 'title',
+    },
+    from_date: {
+      Name: 'from_date',
+      Value: null,
+      data_type: 'date',
+      TableName: 'titles',
+      References: [],
+      IsPrimaryKey: true,
+      IsForeignKey: false,
+      additional_constraints: 'PRIMARY KEY',
+      field_name: 'from_date',
+    },
+    to_date: {
+      Name: 'to_date',
+      Value: null,
+      data_type: 'date',
+      TableName: 'titles',
+      References: [],
+      IsPrimaryKey: false,
+      IsForeignKey: false,
+      additional_constraints: 'NA',
+      field_name: 'to_date',
+    },
+  },
+};
 //
 // State Management for db Schema
 //
@@ -7,14 +339,13 @@ import { devtools, persist } from 'zustand/middleware';
 
 let schemaStore = (set) => ({
   //schemaStore state
-  schemaStore: null,
+  schemaStore: departmentSchema,
   //schema is an object
-  setSchemaStore: (schema) =>
-    set((state) => ({ ...state, schemaStore: schema })),
+  setSchemaStore: (schema) => set((state) => ({ ...state, schemaStore: schema })),
 });
 
 schemaStore = devtools(schemaStore);
-schemaStore = persist(schemaStore);
+// schemaStore = persist(schemaStore);
 const useSchemaStore = create(schemaStore);
 
 export default useSchemaStore;
