+++ src/styles/index.css
@@ -349,8 +349,8 @@
 .table-node label {
   display: block;
   text-align: left;
-  background-color: #2b3a42;
-  color: #f8f4eb;
+  background-color: #075985;
+  color: 7 89 133;
   font-size: 15px;
   font-weight: bold;
   margin-top: -1px;
